+++ packages/graph-explorer/src/core/StateProvider/renderedEntities.test.ts
@@ -4,7 +4,7 @@ import {
   DbState,
   renderHookWithJotai,
 } from "@/utils/testing";
-import { createEdgeId, createVertexId } from "./entityIdType";
+import { createEdgeId, createVertexId } from "@/core/entities";
 import {
   createRenderedEdgeId,
   createRenderedVertexId,
