+++ packages/graph-explorer/src/core/entities/entities.ts
@@ -1,0 +1,7 @@
+import { Vertex, VertexId } from "./vertex";
+import { Edge, EdgeId } from "./edge";
+
+export type Entities = {
+  nodes: Map<VertexId, Vertex>;
+  edges: Map<EdgeId, Edge>;
+};
