+++ packages/graph-explorer/src/core/entities/shared.ts
@@ -1,0 +1,3 @@
+export type EntityRawId = string | number;
+export type EntityPropertyValue = string | number | boolean;
+export type EntityProperties = Record<string, EntityPropertyValue>;
