+++ packages/graph-explorer/src/core/index.ts
@@ -3,8 +3,5 @@ export * from "./ThemeProvider";
 export * from "./featureFlags";
 export * from "./StateProvider";
 export * from "./connector";
-export * from "./createEntities";
-export * from "./entityIdType";
 export * from "./entities";
-export * from "./renderedEntities";
 export * from "./fetchEntityDetails";
